+++ RELEASE_NOTES.md
@@ -2,6 +2,7 @@
 
 ## New in git-machete 3.9.0
 
+- added: `advance` command now also pushes the branch after the merge
 - fixed: `fork-point` no longer specially treats branches merged to its parent
 
 ## New in git-machete 3.8.0